# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/hello/greeting": {
            "post": {
                "description": "Greeting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hello"
                ],
                "summary": "Greeting",
                "operationId": "greeting",
                "parameters": [
                    {
                        "description": "greeting",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.greetingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.response"
                        }
                    }
                }
            }
        },
        "/hello/say": {
            "get": {
                "description": "Show hello",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hello"
                ],
                "summary": "Show hello",
                "operationId": "echo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.response"
                        }
                    }
                }
            }
        },
        "/translation/do-translate": {
            "post": {
                "description": "Translate a text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translation"
                ],
                "summary": "Translate",
                "operationId": "do-translate",
                "parameters": [
                    {
                        "description": "Set up translation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.doTranslateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Translation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.response"
                        }
                    }
                }
            }
        },
        "/translation/history": {
            "get": {
                "description": "Show all translation history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translation"
                ],
                "summary": "Show history",
                "operationId": "history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.historyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Translation": {
            "type": "object",
            "properties": {
                "destination": {
                    "type": "string",
                    "example": "en"
                },
                "original": {
                    "type": "string",
                    "example": "текст для перевода"
                },
                "source": {
                    "type": "string",
                    "example": "auto"
                },
                "translation": {
                    "type": "string",
                    "example": "text for translation"
                }
            }
        },
        "github.com_kuiyonggen_go-clean-template_internal_controller_http_v1.doTranslateRequest": {
            "type": "object",
            "required": [
                "destination",
                "original",
                "source"
            ],
            "properties": {
                "destination": {
                    "type": "string",
                    "example": "en"
                },
                "original": {
                    "type": "string",
                    "example": "текст для перевода"
                },
                "source": {
                    "type": "string",
                    "example": "auto"
                }
            }
        },
        "github.com_kuiyonggen_go-clean-template_internal_controller_http_v1.greetingRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "alice"
                }
            }
        },
        "github.com_kuiyonggen_go-clean-template_internal_controller_http_v1.historyResponse": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Translation"
                    }
                }
            }
        },
        "github.com_kuiyonggen_go-clean-template_internal_controller_http_v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "internal_controller_http_v1.doTranslateRequest": {
            "type": "object",
            "required": [
                "destination",
                "original",
                "source"
            ],
            "properties": {
                "destination": {
                    "type": "string",
                    "example": "en"
                },
                "original": {
                    "type": "string",
                    "example": "текст для перевода"
                },
                "source": {
                    "type": "string",
                    "example": "auto"
                }
            }
        },
        "internal_controller_http_v1.greetingRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "alice"
                }
            }
        },
        "internal_controller_http_v1.historyResponse": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Translation"
                    }
                }
            }
        },
        "internal_controller_http_v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}