# internal/controller/http/v1/translation.go
package v1

import (
	"net/http"

	"github.com/gin-gonic/gin"

	"github.com/kuiyonggen/go-clean-template/internal/entity"
	"github.com/kuiyonggen/go-clean-template/internal/usecase"
	"github.com/kuiyonggen/go-clean-template/internal/usecase/repo"
	"github.com/kuiyonggen/go-clean-template/internal/usecase/webapi"
	"github.com/kuiyonggen/go-clean-template/pkg/logger"
        "github.com/kuiyonggen/go-clean-template/config"
)

type translationRoutes struct {
	t usecase.Translation
	l logger.Interface
}

func newTranslationRoutes(handler *gin.RouterGroup, cfg *config.Config) {
        // Use case
        t := usecase.New(
            repo.New(cfg.Pg),
            webapi.New(),
        )
        r := &translationRoutes{t, cfg.Logger}

	h := handler.Group("/translation")
	{
		h.GET("/history", r.history)
		h.POST("/do-translate", r.doTranslate)
	}
}

type historyResponse struct {
	History []entity.Translation `json:"history"`
}

// @Summary      Show history
// @Description  Show all translation history
// @ID           history
// @Tags               translation
// @Accept       json
// @Produce      json
// @Success      200  {object}  historyResponse
// @Failure      500  {object}  response
// @Router       /translation/history [get]
func (r *translationRoutes) history(c *gin.Context) {
	translations, err := r.t.History(c.Request.Context())
	if err != nil {
		r.l.Error(err, "http - v1 - history")
		errorResponse(c, http.StatusInternalServerError, "database problems")

		return
	}

	c.JSON(http.StatusOK, historyResponse{translations})
}

type doTranslateRequest struct {
	Source      string `json:"source"       binding:"required"  example:"auto"`
	Destination string `json:"destination"  binding:"required"  example:"en"`
	Original    string `json:"original"     binding:"required"  example:"текст для перевода"`
}

// @Summary      Translate
// @Description  Translate a text
// @ID           do-translate
// @Tags               translation
// @Accept       json
// @Produce      json
// @Param        request  body      doTranslateRequest  true  "Set up translation"
// @Success      200      {object}  entity.Translation
// @Failure      400      {object}  response
// @Failure      500      {object}  response
// @Router       /translation/do-translate [post]
func (r *translationRoutes) doTranslate(c *gin.Context) {
	var request doTranslateRequest
	if err := c.ShouldBindJSON(&request); err != nil {
		r.l.Error(err, "http - v1 - doTranslate")
		errorResponse(c, http.StatusBadRequest, "invalid request body")

		return
	}

	translation, err := r.t.Translate(
		c.Request.Context(),
		entity.Translation{
			Source:      request.Source,
			Destination: request.Destination,
			Original:    request.Original,
		},
	)
	if err != nil {
		r.l.Error(err, "http - v1 - doTranslate")
		errorResponse(c, http.StatusInternalServerError, "translation service problems")

		return
	}

	c.JSON(http.StatusOK, translation)
}